+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 ### HEAD
 
-* Updated build to current versions
+* Removed deprecated `window.Defunctr`
+* Removed deprecated `defunctr.detective`
+* NuGet installs to `{content}/scripts/defunctr/` now instead of `{content}/Scripts/Vendor/Defunctr/`
 * Adding document to illustrate how to migrate a Visual Studio project to NPM so we can stop distributing through NuGet
 * Starting to add automated test suite
 
 ### 1.3.1 (August 13, 2017)
 * Fixed issue where Opera 12 through 14 were detecting as Chrome, WebKit, and Opera
-* Added defunctr.browser which will replace defunctr.detective in 2.0
-* Deprecated defunctr.detective
+* Added `defunctr.browser` which will replace `defunctr.detective` in 2.0
+* Deprecated `defunctr.detective`
 * Updated build to current versions
 
 ### 1.3.0 (February 10, 2017)